Restructuring changes
-put together major components to replace mainactivity
-task activity recycler view replaced with a textview
-added option to add a description about the task being completed
-completed tasks can be removed on longpress or tap to view description
-cleaned up code and removed unnecessary files
