--- app/src/main/java/com/example/finalProject/Contact.kt
@@ -1,7 +1,0 @@
-package com.example.finalProject
-
-data class Contact (
-    var id: Int? = null,
-    var name: String? = null,
-    var email: String? = null
-)
