+++ app/src/main/res/layout/activity_completed_task.xml
@@ -11,12 +11,24 @@
         android:layout_width="329dp"
         android:layout_height="508dp"
         android:layout_marginStart="1dp"
-        android:layout_marginTop="1dp"
         android:layout_marginEnd="1dp"
         android:layout_marginBottom="1dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/compledTasks_tv"
+        app:layout_constraintVertical_bias="0.051"
         tools:listitem="@layout/row_item" />
+
+    <TextView
+        android:id="@+id/compledTasks_tv"
+        android:layout_width="355dp"
+        android:layout_height="68dp"
+        android:layout_marginTop="36dp"
+        android:text="These are your completed tasks, tap to view description or hold down on the task if you want to delete it"
+        android:textAlignment="center"
+        android:textSize="18sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 </androidx.constraintlayout.widget.ConstraintLayout>
