--- app/src/main/res/layout/activity_task.xml
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".taskActivity">
-
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/task_recycler_view"
-        android:layout_width="393dp"
-        android:layout_height="503dp"
-        android:layout_marginEnd="8dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/typeTaskBtn"
-        tools:listitem="@layout/row_item" />
-
-    <TextView
-        android:id="@+id/taskTitle"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:text="These are your current tasks, hold down on an item to delete or complete a task!"
-        android:textAlignment="center"
-        android:textSize="18sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <Button
-        android:id="@+id/typeTaskBtn"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="64dp"
-        android:onClick="addTask"
-        android:text="Get task:"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/taskTitle" />
-
-    <Spinner
-        android:id="@+id/categorySpinner"
-        android:layout_width="184dp"
-        android:layout_height="44dp"
-        android:layout_marginTop="9dp"
-        android:layout_marginBottom="15dp"
-        android:entries="@array/spinnerItems"
-        app:layout_constraintBottom_toTopOf="@+id/task_recycler_view"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toEndOf="@+id/typeTaskBtn"
-        app:layout_constraintTop_toBottomOf="@+id/taskTitle"
-        app:layout_constraintVertical_bias="0.409" />
-
-    <Button
-        android:id="@+id/completedTask_Btn"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:onClick="launchCompletedTaskActivity"
-        android:text="View Completed Tasks"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.497"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/task_recycler_view"
-        app:layout_constraintVertical_bias="0.311" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
